I added in application.properties necessary metadata to connect to postgres DB.
I modified class Student, so now it creates a table student in Student DB.

I am scared because it is a little sick.
I had to have a little "fun" with terminal so type to me if the program will not work. I bet it won't because you don't have a DB.